feat: Create Core Token Model
