+++ data/core/src/main/java/com/seugi/data/core/model/TokenModel.kt
@@ -1,0 +1,13 @@
+package com.seugi.data.core.model
+
+
+data class TokenModel(
+    val accessToken: String,
+    val refreshToken: String
+)
+
+
+data class LocalTokenModel(
+    val accessToken: String?,
+    val refreshToken: String?
+)
