+++ data/member/src/main/java/com/seugi/data/member/MemberRepository.kt
@@ -1,12 +1,12 @@
 package com.seugi.data.member
 
 import com.seugi.common.model.Result
-import com.seugi.data.member.model.EmailSignInModel
+import com.seugi.data.core.model.TokenModel
 import com.seugi.network.request.EmailSignInRequest
 import kotlinx.coroutines.flow.Flow
 
 interface MemberRepository {
-    suspend fun emailSignIn(body: EmailSignInRequest): Flow<Result<EmailSignInModel>>
+    suspend fun emailSignIn(body: EmailSignInRequest): Flow<Result<TokenModel>>
     suspend fun getCode(email: String): Flow<Result<Int>>
 
     suspend fun emailSignUp(name: String, email: String, password: String, code: String): Flow<Result<Int>>
