--- data/member/src/main/java/com/seugi/data/member/model/EmailSignInModel.kt
@@ -1,6 +1,0 @@
-package com.seugi.data.member.model
-
-data class EmailSignInModel(
-    val accessToken: String,
-    val refreshToken: String,
-)
