+++ data/oauth/src/main/java/com/seugi/data/oauth/OauthRepository.kt
@@ -2,6 +2,7 @@ package com.seugi.data.oauth
 
 import kotlinx.coroutines.flow.Flow
 import com.seugi.common.model.Result
+import com.seugi.data.core.model.TokenModel
 
 interface OauthRepository {
