--- data/oauth/src/main/java/com/seugi/data/oauth/TokenModel.kt
@@ -1,14 +1,0 @@
-package com.seugi.data.oauth
-
-import com.seugi.network.core.response.TokenResponse
-
-data class TokenModel(
-    val accessToken: String,
-    val refreshToken: String,
-)
-
-
-fun TokenResponse.toModel(): TokenModel = TokenModel(
-    accessToken = this.accessToken,
-    refreshToken = this.refreshToken
-)
