+++ data/oauth/src/main/java/com/seugi/data/oauth/repository/OauthRepositoryImpl.kt
@@ -4,9 +4,9 @@ import com.seugi.common.model.Result
 import com.seugi.common.model.asResult
 import com.seugi.common.utiles.DispatcherType
 import com.seugi.common.utiles.SeugiDispatcher
+import com.seugi.data.core.mapper.toModel
+import com.seugi.data.core.model.TokenModel
 import com.seugi.data.oauth.OauthRepository
-import com.seugi.data.oauth.TokenModel
-import com.seugi.data.oauth.toModel
 import com.seugi.network.core.response.safeResponse
 import com.seugi.network.oauth.OauthDatasource
 import kotlinx.coroutines.CoroutineDispatcher
