--- data/token/src/main/java/com/seugi/data/token/mapper/TokenMapper.kt
@@ -1,9 +1,0 @@
-package com.seugi.data.token.mapper
-
-import com.seugi.data.token.model.TokenModel
-import com.seugi.local.room.model.TokenEntity
-
-fun TokenEntity?.toModel() = TokenModel(
-    accessToken = this?.token,
-    refreshToken = this?.refreshToken,
-)
