--- data/token/src/main/java/com/seugi/data/token/model/TokenModel.kt
@@ -1,6 +1,0 @@
-package com.seugi.data.token.model
-
-data class TokenModel(
-    val accessToken: String?,
-    val refreshToken: String?,
-)
