+++ data/token/src/main/java/com/seugi/data/token/repository/TokenRepositoryImpl.kt
@@ -4,9 +4,9 @@ import com.seugi.common.model.Result
 import com.seugi.common.model.asResult
 import com.seugi.common.utiles.DispatcherType
 import com.seugi.common.utiles.SeugiDispatcher
+import com.seugi.data.core.mapper.toModel
+import com.seugi.data.core.model.LocalTokenModel
 import com.seugi.data.token.TokenRepository
-import com.seugi.data.token.mapper.toModel
-import com.seugi.data.token.model.TokenModel
 import com.seugi.local.room.dao.TokenDao
 import com.seugi.local.room.model.TokenEntity
 import com.seugi.network.token.TokenDatasource
@@ -30,7 +30,7 @@ class TokenRepositoryImpl @Inject constructor(
         )
     }
 
-    override suspend fun getToken(): Flow<Result<TokenModel>> {
+    override suspend fun getToken(): Flow<Result<LocalTokenModel>> {
         return flow {
             val tokenEntity = tokenDao.getToken()
             emit(tokenEntity.toModel())
@@ -39,7 +39,7 @@ class TokenRepositoryImpl @Inject constructor(
             .asResult()
     }
 
-    override suspend fun newToken(): Flow<Result<TokenModel>> {
+    override suspend fun newToken(): Flow<Result<LocalTokenModel>> {
         return flow {
             tokenDao.getToken()?.apply {
                 val newTokenEntity = TokenEntity(
