+++ network/sign-in/src/main/java/com/seugi/network/MemberDatasource.kt
@@ -2,11 +2,11 @@ package com.seugi.network
 
 import com.seugi.network.core.response.BaseResponse
 import com.seugi.network.core.response.Response
+import com.seugi.network.core.response.TokenResponse
 import com.seugi.network.request.EmailSignInRequest
-import com.seugi.network.response.EmailSignInResponse
 
 interface MemberDatasource {
-    suspend fun emailSignIn(body: EmailSignInRequest): BaseResponse<EmailSignInResponse>
+    suspend fun emailSignIn(body: EmailSignInRequest): BaseResponse<TokenResponse>
 
     suspend fun getCode(email: String): Response
     suspend fun emailSignUp(name: String, email: String, password: String, code: String): Response
