--- network/sign-in/src/main/java/com/seugi/network/response/EmailSignInResponse.kt
@@ -1,9 +1,0 @@
-package com.seugi.network.response
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class EmailSignInResponse(
-    val accessToken: String,
-    val refreshToken: String,
-)
